feat #88 ChatRoom Datasource 소스 코드 내 데이터 모델 추가, chat_message_mapper, chat_room_message_res 삭제
